Also deserialize tool calls in in multi-turn env
